refact!: move from using IntelliJ to maven
This is a WIP commit where we do the initial process of removing IntelliJ
as our project management tool to instead use Maven.

The goal is to allow users of different IDEs to edit the project
and make the project available through Maven. This will allow other
open source projects to more easily extend or use this project; more
specifically we intend to redo how the pump plugin files are created.

For the moment this is a breaking change since it doesn't yet provide an
alternative method to generate the required binaries. This will come next.

Side Effects:
1 - Removed binary copies of libraries we depend on
2 - Removed from git all IntelliJ specific files
3 - Removed packaged binaries
4 - Renamed software submodules to remove the redundant names.
    For example:
      A - Applications/pumpControl_IJ became Applications/ImageJ
      B - Pumps/pumpControl.harvardElite became PumpControl/HarvardElite
5 - The project doesn't yet provide an alternative to generating binaries.