refactor: remove reflection-based jar loading
It seems that at the time I wrote this, I either didn't understand the JVM class loader, or thought doing it this way was "coool".

I now both understand the JVM and think this is hacky (plus reflection is a no no).

So I've removed the hard-coded jar loader and added an example bat file (that will likely be used later for packaged releases).