fix: micro-manager not closing
There was some oddity in trying to load the jars in plugins\Micro-Manager that meant that, after closing our software, the micro-manager window would refuse to close (something about X class not implementing Y interface).

The solution was simple, to place the required libraries in the root Plugins folder instead.